requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-wtf>=1.2
wtforms>=3.1
python-dotenv>=1.0
pytest>=7.0
Pillow>=9.0
pytesseract>=0.3
# Flask-Migrate removed for minimal local setup
